# AUTOMATION/web/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulglade — API du studio",
    "description": "\nBackend local du studio Soulglade. Aucune authentification : le serveur\nn'accepte que ce qui vient de lui-même (voir `api/security.py`).\n\n**Conventions transverses**\n\n* presque toute route lit `?character=<id>` et le valide avant de toucher au\n  disque ; `/img` l'exige (pas de repli sur un personnage par défaut) ;\n* toute réponse porte un corps JSON, succès comme échec. Une erreur a la forme\n  `{\"ok\": false, \"erreur\": \"<texte français destiné à l'écran>\"}` ;\n* les uploads voyagent en base64 dans un corps JSON, jamais en\n  `multipart/form-data` — conséquence voulue du garde d'origine.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/scenes": {
      "get": {
        "summary": "Banque de scènes du personnage",
        "description": "The bank, plus everything the Créer screen's cards need in ONE call.",
        "operationId": "get_scene_bank_api_scenes_get",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneBankResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enregistrer la banque de scènes",
        "description": "Writes scenes.json after validation, with a .bak rotation over three\ngenerations.\n\nThe server no longer trusts the frontend on the shape of the bank: that is\nthe check that was missing on 25/08/2026 when an interface-side rebuild\nwiped the creative journey from all 16 scenes with nothing to stop it. It\nwrites a file `build_jobs` will be able to read, or it refuses.",
        "operationId": "save_scene_bank_api_scenes_post",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneBankSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Banque refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneBankRejected"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/creative": {
      "get": {
        "summary": "Taxonomie du parcours créatif",
        "description": "Journey taxonomy: intentions, tones, intensity scale.",
        "operationId": "get_creative_taxonomy_api_creative_get",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreativeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compose": {
      "post": {
        "summary": "Composer des scènes depuis une intention en français",
        "description": "Turns a French intention into scenes ready to be reviewed.\n\nGoes through the local LLM served by ComfyUI, in an executor: `composer`\ntalks to it with blocking urllib and a /history poll, exactly like the\nproduction runner. Nothing is written to the bank here — the proposal comes\nback for review, and the user saves it through POST /api/scenes.",
        "operationId": "compose_scenes_api_compose_post",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComposeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComposeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Le composeur a échoué"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "erreur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erreur"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ActionResponse",
        "description": "The universal answer of an action route, `{ok: bool, erreur?: str}`.\n\nAUDIT §5.4: the convention holds on 4xx and 5xx too — `api.js` reads a JSON\nbody on every response, whatever the status.\n\nALWAYS DECLARE IT WITH `response_model_exclude_unset=True`. The old handlers\nreturned `{\"ok\": true}` and nothing else on success; a plain response_model\nwould add `\"erreur\": null` to every one of them. Harmless to `erreurDe()`,\nwhich only tests `ok === false` — but this migration is meant to keep the\nbodies identical, and a shape that drifts once drifts twice."
      },
...
      "ComposeRequest": {
        "properties": {
          "intention": {
            "type": "string",
            "title": "Intention",
            "default": ""
          },
          "intention_cible": {
            "type": "string",
            "title": "Intention Cible",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ComposeRequest",
        "description": "`intention` is the free French text describing what is wanted;\n`intention_cible` is the taxonomy KEY being imposed. Confusing the two put\nthe French sentence into the scenes' intention field. `category` is the old\nname of `intention_cible`, still accepted."
      },
      "ComposeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "scenes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Scenes"
          },
          "brut": {
            "type": "string",
            "title": "Brut"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "scenes",
          "brut"
        ],
        "title": "ComposeResponse",
        "description": "Proposed scenes, ready to be reviewed — never written to the bank on\ntheir own. `brut` is the LLM's raw answer, truncated to 2000 characters, so\na disappointing proposal can be diagnosed."
      },
...
      "CreativeIntention": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "key"
        ],
        "title": "CreativeIntention"
      },
      "CreativeResponse": {
        "properties": {
          "intentions": {
            "items": {
              "$ref": "#/components/schemas/CreativeIntention"
            },
            "type": "array",
            "title": "Intentions"
          },
          "tones": {
            "items": {
              "$ref": "#/components/schemas/CreativeTone"
            },
            "type": "array",
            "title": "Tones"
          },
          "intensity": {
            "items": {
              "$ref": "#/components/schemas/IntensityTier"
            },
            "type": "array",
            "title": "Intensity"
          },
          "niveau_natif": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niveau Natif"
          }
        },
        "type": "object",
        "required": [
          "intentions",
          "tones",
          "intensity"
        ],
        "title": "CreativeResponse"
      },
      "CreativeTone": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "expression": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpressionRangeParams"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt_add": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Add"
          },
          "couche": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "monde",
                  "surcharge",
                  "personnage"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Couche"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "key"
        ],
        "title": "CreativeTone"
      },
...
      "ErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ErrorResponse",
        "description": "What every rejection looks like. `erreur` is French: it is displayed to\nthe user verbatim (AUDIT §5.5)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneBankRejected": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          },
          "problemes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problemes"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "SceneBankRejected",
        "description": "400 of a refused save. `erreur` is the first problem — what the screen\nshows; `problemes` is the whole list, for the details panel."
      },
      "SceneBankResponse": {
        "properties": {
          "data": {
            "title": "Data"
          },
          "categories": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Categories"
          },
          "scene_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scene Ids"
          },
          "previews": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ScenePreview"
            },
            "type": "object",
            "title": "Previews"
          },
          "meta": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SceneMeta"
            },
            "type": "object",
            "title": "Meta"
          },
          "stats": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SceneStats"
            },
            "type": "object",
            "title": "Stats"
          },
          "avg_duration": {
            "type": "integer",
            "title": "Avg Duration"
          },
          "poses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Poses"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "data",
          "categories",
          "scene_ids",
          "previews",
          "meta",
          "stats",
          "avg_duration",
          "poses"
        ],
        "title": "SceneBankResponse",
        "description": "`data` is scenes.json VERBATIM — the file belongs to the character, not\nto this layer, so it is relayed untouched and unmodelled."
      },
      "SceneBankSaveRequest": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "data": {
            "title": "Data"
          },
          "autoriser_pertes": {
            "type": "boolean",
            "title": "Autoriser Pertes",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SceneBankSaveRequest",
        "description": "Two accepted forms, `text` (a JSON string) or `data` (the object), and a\ndeliberate escape hatch.\n\n`autoriser_pertes` disarms the batch-erasure guard for ONE save. It is the\npayload key, French, because it is part of the wire contract — renaming it\nwould silently disable the guard for any client still sending the old name.\nNothing in the frontend sends it today; it exists for a deliberate mass\nedit made by hand."
      },
      "SceneMeta": {
        "properties": {
          "intention": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intention"
          },
          "band": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Band"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "tones": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tones"
          },
          "pose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pose"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "band"
        ],
        "title": "SceneMeta",
        "description": "What the SERVER computes about a scene so the frontend does not have to\nreimplement the runner's compatibility defaults (`scene_intention`,\n`scene_band`). Not the scene itself — that travels raw in `data`."
      },
      "ScenePreview": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "name"
        ],
        "title": "ScenePreview",
        "description": "Last image produced BY THIS CHARACTER for that scene, to illustrate the\nscene cards. Without the character, the Créer screen illustrated every\ncard with Léna's images whatever character was open."
      },
      "SceneStats": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "ok": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "n"
        ],
        "title": "SceneStats",
        "description": "Per scene: how many images were produced, how many the HUMAN validated,\nand the mean identity score. `ok` counts the human sort, not the QC verdict\n(base.stats_par_scene). `avg` is null while no image carries a score."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}